app/release-model-service/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
psycopg2==2.9.9
pytest ~= 5.4
Werkzeug==3.0.2
google-auth==2.29.0
google-cloud-bigquery==3.20.1